set `shape_list` to `float32`

+++ mindocr/data/transforms/det_transforms.py
@@ -446,7 +446,7 @@ class DetResize(object):
 
         if 'shape_list' not in data:
             src_h, src_w = data.get('raw_img_shape', (h, w))
-            data['shape_list'] = np.array([src_h, src_w, scale_h, scale_w])
+            data['shape_list'] = np.array([src_h, src_w, scale_h, scale_w], dtype=np.float32)
         else:
             data['shape_list'][2] = data['shape_list'][2] * scale_h
             data['shape_list'][3] = data['shape_list'][3] * scale_h
